feat: refactor PaymentCalculator to use new components for configuration, summary, and installment table
- Introduced `CalcConfiguration`, `AmountResume`, and `InstallmentTable` components for better modularity.
- Removed inline logic for shipping and installment calculations, delegating to new components.
- Updated footer component for consistent layout.
- Enhanced code readability and maintainability.
